lib/js/datachannel_test_2peer/datachanneldict/012.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - DataChannelDict Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label1 = "test";
            var label2 = "tset";
            var expected = "a";
            var dataChannelOptions1 = {
                id : 2,
                ordered : false,
                negotiated : true
            };
            var dataChannelOptions2 = {
                id : 2,
                ordered : true,
                negotiated : true
            };
            var test = async_test("Try to set up a DataChannel by creating two corresponding DataChannels with same id and negotiated true - try to establish a connection with asymmetric options (label, ordered) - send test message");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label1, dataChannelOptions1);
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.readyState, "open", "Wrong readyState");
                        assert_not_equals(dataChannel.ordered, dataChannelOptions2.ordered, "not set ordered different in both channels");
                        assert_not_equals(dataChannel.label, label2, "not set label different in both channels");
                        try {
                            dataChannel.send(expected);
                        } catch(e) {
                            assert_unreached(e.name + ": " + e.message);
                        }
                    });
                    dataChannel.onmessage = test.step_func(function(e) {
                        assert_equals(e.data, expected, "Received wrong data ");
                        test.done();
                    });
                });
            } else {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label2, dataChannelOptions2);
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.readyState, "open", "Wrong readyState");
                        assert_not_equals(dataChannel.ordered, dataChannelOptions1.ordered, "not set ordered different in both channels");
                        assert_not_equals(dataChannel.label, label1, "not set label different in both channels");
                        try {
                            dataChannel.send(expected);
                        } catch(e) {
                            assert_unreached(e.name + ": " + e.message);
                        }
                    });
                    dataChannel.onmessage = test.step_func(function(e) {
                        assert_equals(e.data, expected, "Received wrong data ");
                        test.done();
                    });
                });
            }
		</script>
	</body>
</html>
